requirements.txt
PyYAML~=6.0
Telethon~=1.28.5
cryptg
pillow
aiohttp
sqlalchemy
mysql-connector-python